add admin apis and authorization funtionality

+++ src/app/api/v1/host_apis/[category]/route.ts
@@ -1,20 +1,25 @@
-import { serverError } from "@/BackendFiles/OnError"
+import { serverError, unathorizeError } from "@/BackendFiles/OnError"
 import { getCollection } from "@/BackendFiles/Utils"
 import { connectDB } from "@/BackendFiles/Utils/MongoDB-Utils"
+import { verifyHost } from "@/BackendFiles/Utils/auth-utils"
 import { NextRequest, NextResponse } from "next/server"
 
 export const POST = async (req: NextRequest, { params }: any) => {
     try {
-        const { category } = params
-        const data = await req.json()
-        const dataObtainCollections = getCollection(req, params)
-        await connectDB()
-        const newData = new dataObtainCollections(data)
-        const result = await newData.save()
-        if (result._id) {
-            return NextResponse.json({ message: `${category.slice(0, category.length - 1)} creat successful` })
+        const isHost = await verifyHost(req)
+        if (isHost) {
+            const { category } = params
+            const data = await req.json()
+            const dataObtainCollections = getCollection(req, params)
+            await connectDB()
+            const newData = new dataObtainCollections(data)
+            const result = await newData.save()
+            if (result._id) {
+                return NextResponse.json({ message: `${category.slice(0, category.length - 1)} creat successful` })
+            }
+            return serverError(req)
         }
-        return serverError(req)
+        return unathorizeError(req)
     } catch (error: any) {
         console.log(error.message)
         return serverError(req)
@@ -23,22 +28,26 @@ export const POST = async (req: NextRequest, { params }: any) => {
 
 export const PATCH = async (req: NextRequest, { params }: any) => {
     try {
-        const { category } = params
-        const { searchParams } = new URL(req.url)
-        const id = searchParams.get("id")
-        const data = await req.json()
-        const updateDoc = {
-            $set: {
-                ...data
+        const isHost = await verifyHost(req)
+        if (isHost) {
+            const { category } = params
+            const { searchParams } = new URL(req.url)
+            const id = searchParams.get("id")
+            const data = await req.json()
+            const updateDoc = {
+                $set: {
+                    ...data
+                }
             }
+            const dataObtainCollections = getCollection(req, params)
+            await connectDB()
+            const result = await dataObtainCollections.findOneAndUpdate({ _id: id }, updateDoc)
+            if (result._id) {
+                return NextResponse.json({ message: `${category.slice(0, category.length - 1)} update successful` })
+            }
+            return serverError(req)
         }
-        const dataObtainCollections = getCollection(req, params)
-        await connectDB()
-        const result = await dataObtainCollections.findOneAndUpdate({ _id: id }, updateDoc)
-        if (result._id) {
-            return NextResponse.json({ message: `${category.slice(0, category.length - 1)} update successful` })
-        }
-        return serverError(req)
+        return unathorizeError(req)
     } catch (error) {
         return serverError(req)
     }
@@ -46,17 +55,21 @@ export const PATCH = async (req: NextRequest, { params }: any) => {
 
 export const DELETE = async (req: NextRequest, { params }: any) => {
     try {
-        const { category } = params
-        const { searchParams } = new URL(req.url)
-        const id = searchParams.get("id")
-        const dataObtainCollections = getCollection(req, params)
-        await connectDB()
-        const result = await dataObtainCollections.deleteOne({ _id: id })
-        console.log(result)
-        if (result.deletedCount > 0) {
-            return NextResponse.json({ message: `${category.slice(0, category.length - 1)} update successful` })
+        const isHost = await verifyHost(req)
+        if (isHost) {
+            const { category } = params
+            const { searchParams } = new URL(req.url)
+            const id = searchParams.get("id")
+            const dataObtainCollections = getCollection(req, params)
+            await connectDB()
+            const result = await dataObtainCollections.deleteOne({ _id: id })
+            console.log(result)
+            if (result.deletedCount > 0) {
+                return NextResponse.json({ message: `${category.slice(0, category.length - 1)} update successful` })
+            }
+            return serverError(req)
         }
-        return serverError(req)
+        return unathorizeError(req)
     } catch (error) {
         return serverError(req)
     }
